requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.1
asn1crypto==1.5.1
boto3==1.17.18
botocore==1.20.18
certifi==2020.12.5
cffi==1.15.1
chardet==4.0.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.2
cssselect2==0.6.0
dj-database-url==0.5.0
Django==3.1.5
django-cors-headers==3.7.0
django-storages==1.11.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.5.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==2.10
jmespath==0.10.0
lxml==4.9.0
oscrypto==1.3.0
Pillow==9.1.0
psycopg2-binary==2.9.3
pycodestyle==2.7.0
pycparser==2.21
pyHanko==0.13.1
pyhanko-certvalidator==0.19.5
PyJWT==1.7.1
PyPDF2==1.26.0
PyPDF3==1.0.6
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2020.5
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.10
requests==2.25.1
s3transfer==0.3.7
six==1.15.0
sqlparse==0.4.1
svglib==1.3.0
tinycss2==1.1.1
toml==0.10.2
tqdm==4.64.0
typing-extensions==3.7.4.3
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.3
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.8
zipp==3.4.0
